+++ joulescope_ui/help_ui.py
@@ -38,7 +38,7 @@ _MD_HEADER = """\
 <head>
   <meta charset="utf-8">
   <meta http-equiv="X-UA-Compatible" content="IE=edge,chrome=1">
-  <title>Joulescope Credits</title>
+  <title>Changelog</title>
   {style}
 </head>